% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SumOfSquaresOfDeviation.R
\name{SSDM}
\alias{SSDM}
\title{Provides function to find sum of squares of deviation from mean}
\usage{
SSDM(vector1)
}
\arguments{
\item{vector1}{A distribution vector.}
}
\value{
z A value.
}
\description{
A sum of squares of deviation - scalar
}
